+++ Changelog.md
@@ -7,6 +7,7 @@
 - New: Item tag "securitycraft:reinforced/glass_panes"
 - Change: The recipe for the Laser Block now uses 4 glass panes in total
 - Change: You can no longer break the code of passcode blocks/briefcases that you own
+- Change: Universal Block Reinforcers can no longer be enchanted with Mending. Old reinforcers lose their Mending enchantment when used. This is to make it more enticing to upgrade to a level 3 reinforcer. In the future, this idea will be expanded upon.
 - Fix: Crash when opening the customization GUI of a block that has no options (like the Projector)
 - Fix: Blocks can be pushed into a Sentry using a piston
 - Fix: The block at x=0, y=0, z=0 gets removed when placing a Sentry